Add an env.path file to the php-composter directory
Closes #10

This adds an env.path file to .git/php-composter/ directory with the one line content of what the composer vendor directory is.

Later we check the file's existence, and then `cat` contents and use that as a bases to where our vendor directory is, and thus, the entry autoload file.